feat: integrate actions metrics and replace billing collectors
