+++ changelog/unreleased/action-metrics.md
@@ -1,0 +1,6 @@
+Enhancement: Add metrics for GitHub actions
+
+We've added new metrics for the observability of GitHub actions, they are
+disabled by default because it could result in a high cardinality of the labels.
+
+https://github.com/promhippie/github_exporter/issues/123
