+++ changelog/unreleased/update-deps.md
@@ -1,0 +1,6 @@
+Enhancement: Update all releated dependencies
+
+We've updated all dependencies to the latest available versions, including the
+build tools provided by Bingo.
+
+https://github.com/promhippie/github_exporter/pull/174
