+++ docs/partials/envvars.md
@@ -52,11 +52,8 @@ GITHUB_EXPORTER_COLLECTOR_ORGS
 GITHUB_EXPORTER_COLLECTOR_REPOS
 : Enable collector for repos, defaults to `true`
 
+GITHUB_EXPORTER_COLLECTOR_BILLING
+: Enable collector for billing, defaults to `false`
+
 GITHUB_EXPORTER_COLLECTOR_ACTIONS
 : Enable collector for actions, defaults to `false`
-
-GITHUB_EXPORTER_COLLECTOR_PACKAGES
-: Enable collector for packages, defaults to `false`
-
-GITHUB_EXPORTER_COLLECTOR_STORAGE
-: Enable collector for storage, defaults to `false`
