+++ pkg/config/config.go
@@ -36,11 +36,10 @@ type Target struct {
 
 // Collector defines the collector specific configuration.
 type Collector struct {
-	Orgs     bool
-	Repos    bool
-	Actions  bool
-	Packages bool
-	Storage  bool
+	Orgs    bool
+	Repos   bool
+	Billing bool
+	Actions bool
 }
 
 // Config is a combination of all available configurations.
