+++ pkg/exporter/action.go
@@ -1,12 +1,7 @@
 package exporter
 
 import (
-	"context"
-	"fmt"
-	"time"
-
 	"github.com/go-kit/log"
-	"github.com/go-kit/log/level"
 	"github.com/google/go-github/v50/github"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/promhippie/github_exporter/pkg/config"
@@ -20,10 +15,8 @@ type ActionCollector struct {
 	duration *prometheus.HistogramVec
 	config   config.Target
 
-	MinutesUsed          *prometheus.Desc
-	MinutesUsedBreakdown *prometheus.Desc
-	PaidMinutesUsed      *prometheus.Desc
-	IncludedMinutes      *prometheus.Desc
+	WorkflowCount    *prometheus.Desc
+	WorkflowDuration *prometheus.Desc
 }
 
 // NewActionCollector returns a new ActionCollector.
@@ -32,7 +25,7 @@ func NewActionCollector(logger log.Logger, client *github.Client, failures *prom
 		failures.WithLabelValues("action").Add(0)
 	}
 
-	labels := []string{"type", "name"}
+	labels := []string{"org", "repo", "event", "name", "job", "status", "head_branch", "runner_node_arch", "runner_node_os", "runner_node_type", "retry"}
 	return &ActionCollector{
 		client:   client,
 		logger:   log.With(logger, "collector", "action"),
@@ -40,27 +33,27 @@ func NewActionCollector(logger log.Logger, client *github.Client, failures *prom
 		duration: duration,
 		config:   cfg,
 
-		MinutesUsed: prometheus.NewDesc(
-			"github_action_billing_minutes_used",
-			"Total action minutes used for this type",
+		// github_runner_status
+		// []string{"repo", "os", "name", "id", "busy"},
+		// https://github.com/Spendesk/github-actions-exporter/blob/develop/pkg/metrics/get_runners_from_github.go
+
+		// github_runner_organization_status
+		// []string{"organization", "os", "name", "id", "busy"},
+		// https://github.com/Spendesk/github-actions-exporter/blob/develop/pkg/metrics/get_runners_organization_from_github.go
+
+		// github_runner_enterprise_status
+		// []string{"os", "name", "id"}
+		// https://github.com/Spendesk/github-actions-exporter/blob/develop/pkg/metrics/get_runners_enterprise_from_github.go
+
+		WorkflowCount: prometheus.NewDesc(
+			"github_action_workflow_count",
+			"Number of workflow runs",
 			labels,
 			nil,
 		),
-		MinutesUsedBreakdown: prometheus.NewDesc(
-			"github_action_billing_minutes_used_breakdown",
-			"Total action minutes used for this type broken down by operating system",
-			append(labels, "os"),
-			nil,
-		),
-		PaidMinutesUsed: prometheus.NewDesc(
-			"github_action_billing_paid_minutes",
-			"Total paid minutes used for this type",
-			labels,
-			nil,
-		),
-		IncludedMinutes: prometheus.NewDesc(
-			"github_action_billing_included_minutes",
-			"Included minutes for this type",
+		WorkflowDuration: prometheus.NewDesc(
+			"github_action_workflow_duration_ms",
+			"Duration of workflow runs",
 			labels,
 			nil,
 		),
@@ -70,177 +63,18 @@ func NewActionCollector(logger log.Logger, client *github.Client, failures *prom
 // Metrics simply returns the list metric descriptors for generating a documentation.
 func (c *ActionCollector) Metrics() []*prometheus.Desc {
 	return []*prometheus.Desc{
-		c.MinutesUsed,
-		c.MinutesUsedBreakdown,
-		c.PaidMinutesUsed,
-		c.IncludedMinutes,
+		c.WorkflowCount,
+		c.WorkflowDuration,
 	}
 }
 
 // Describe sends the super-set of all possible descriptors of metrics collected by this Collector.
 func (c *ActionCollector) Describe(ch chan<- *prometheus.Desc) {
-	ch <- c.MinutesUsed
-	ch <- c.MinutesUsedBreakdown
-	ch <- c.PaidMinutesUsed
-	ch <- c.IncludedMinutes
+	ch <- c.WorkflowCount
+	ch <- c.WorkflowDuration
 }
 
 // Collect is called by the Prometheus registry when collecting metrics.
-func (c *ActionCollector) Collect(ch chan<- prometheus.Metric) {
-	for _, name := range c.config.Enterprises.Value() {
-		ctx, cancel := context.WithTimeout(context.Background(), c.config.Timeout)
-		defer cancel()
+func (c *ActionCollector) Collect(_ chan<- prometheus.Metric) {
 
-		req, err := c.client.NewRequest(
-			"GET",
-			fmt.Sprintf("/enterprises/%s/settings/billing/actions", name),
-			nil,
-		)
-
-		if err != nil {
-			level.Error(c.logger).Log(
-				"msg", "Failed to prepare request",
-				"type", "enterprise",
-				"name", name,
-				"err", err,
-			)
-
-			c.failures.WithLabelValues("action").Inc()
-			continue
-		}
-
-		record := &actionReponse{}
-		now := time.Now()
-		_, err = c.client.Do(ctx, req, record)
-		c.duration.WithLabelValues("action").Observe(time.Since(now).Seconds())
-
-		if err != nil {
-			level.Error(c.logger).Log(
-				"msg", "Failed to fetch billing",
-				"type", "enterprise",
-				"name", name,
-				"err", err,
-			)
-
-			c.failures.WithLabelValues("action").Inc()
-			continue
-		}
-
-		labels := []string{
-			"enterprise",
-			name,
-		}
-
-		ch <- prometheus.MustNewConstMetric(
-			c.MinutesUsed,
-			prometheus.GaugeValue,
-			float64(record.TotalMinutesUsed),
-			labels...,
-		)
-
-		ch <- prometheus.MustNewConstMetric(
-			c.PaidMinutesUsed,
-			prometheus.GaugeValue,
-			float64(record.TotalPaidMinutesUsed),
-			labels...,
-		)
-
-		ch <- prometheus.MustNewConstMetric(
-			c.IncludedMinutes,
-			prometheus.GaugeValue,
-			float64(record.IncludedMinutes),
-			labels...,
-		)
-
-		for os, value := range record.MinutesUsedBreakdown {
-			ch <- prometheus.MustNewConstMetric(
-				c.MinutesUsedBreakdown,
-				prometheus.GaugeValue,
-				float64(value),
-				append(labels, os)...,
-			)
-		}
-	}
-
-	for _, name := range c.config.Orgs.Value() {
-		ctx, cancel := context.WithTimeout(context.Background(), c.config.Timeout)
-		defer cancel()
-
-		req, err := c.client.NewRequest(
-			"GET",
-			fmt.Sprintf("/orgs/%s/settings/billing/actions", name),
-			nil,
-		)
-
-		if err != nil {
-			level.Error(c.logger).Log(
-				"msg", "Failed to prepare request",
-				"type", "org",
-				"name", name,
-				"err", err,
-			)
-
-			c.failures.WithLabelValues("action").Inc()
-			continue
-		}
-
-		record := &actionReponse{}
-		now := time.Now()
-		_, err = c.client.Do(ctx, req, record)
-		c.duration.WithLabelValues("action").Observe(time.Since(now).Seconds())
-
-		if err != nil {
-			level.Error(c.logger).Log(
-				"msg", "Failed to fetch billing",
-				"type", "org",
-				"name", name,
-				"err", err,
-			)
-
-			c.failures.WithLabelValues("action").Inc()
-			continue
-		}
-
-		labels := []string{
-			"org",
-			name,
-		}
-
-		ch <- prometheus.MustNewConstMetric(
-			c.MinutesUsed,
-			prometheus.GaugeValue,
-			record.TotalMinutesUsed,
-			labels...,
-		)
-
-		ch <- prometheus.MustNewConstMetric(
-			c.PaidMinutesUsed,
-			prometheus.GaugeValue,
-			record.TotalPaidMinutesUsed,
-			labels...,
-		)
-
-		ch <- prometheus.MustNewConstMetric(
-			c.IncludedMinutes,
-			prometheus.GaugeValue,
-			record.IncludedMinutes,
-			labels...,
-		)
-
-		for os, value := range record.MinutesUsedBreakdown {
-			ch <- prometheus.MustNewConstMetric(
-				c.MinutesUsedBreakdown,
-				prometheus.GaugeValue,
-				value,
-				append(labels, os)...,
-			)
-		}
-	}
-}
-
-type actionReponse struct {
-	TotalMinutesUsed     float64            `json:"total_minutes_used"`
-	TotalPaidMinutesUsed float64            `json:"total_paid_minutes_used"`
-	IncludedMinutes      float64            `json:"included_minutes"`
-	MinutesUsedBreakdown map[string]float64 `json:"minutes_used_breakdown"`
 }
